requirements.txt
python-pptx==0.6.21
regex==2022.3.15
streamlit==1.22.0
streamlit-aggrid==0.3.3
streamlit-folium==0.6.15
openai==0.27.2
OpenAIAuth==0.3.6
openapi-schema-pydantic==1.2.4